refactor: 플레이어의 승리 판단 로직을 Player -> MatchResultType으로 이동
