+++ src/test/kotlin/blackjack/domain/gamer/PlayerTest.kt
@@ -50,47 +50,4 @@ class PlayerTest : StringSpec({
         player.init(InitCard(cards))
         player.captureCards() shouldBe PlayerCards(name, cards)
     }
-
-    "플레이어가 버스트 상태면 딜러가 버스트 상태여도 패배한다" {
-        val dealer = Dealer()
-        dealer.init(initCard(heartQueen(), heartTwo())) // q(10) + 2 = 12
-        dealer.hit(heartKing()) // 12 + k(10) = 22
-        val player = player("test")
-        player.init(initCard(spadeTen(), spadeFour())) // 10 + 4 = 14
-        player.hit(spadeKing()) // 14 + 10 = 24
-        player.match(dealer).matchResultType shouldBe MatchResultType.LOSE
-    }
-
-    "딜러가 버스트 상태라면 플레이어의 점수가 낮더라도 승리한다" {
-        val dealer = Dealer()
-        dealer.init(initCard(heartQueen(), heartTwo())) // q(10) + 2 = 12
-        dealer.hit(heartKing()) // 12 + k(10) = 22
-        val player = player("test")
-        player.init(initCard(spadeTwo(), spadeFour()))
-        player.match(dealer).matchResultType shouldBe MatchResultType.WIN
-    }
-
-    "플레이어의 점수가 딜러의 점수보다 높다면 승리한다" {
-        val dealer = Dealer()
-        dealer.init(initCard(heartQueen(), heartTwo())) // q(10) + 2 = 12
-        val player = player("test")
-        player.init(initCard(spadeKing(), spadeFour())) // k(10) + 4 = 14
-        player.match(dealer).matchResultType shouldBe MatchResultType.WIN
-    }
-
-    "플레이어의 점수가 딜러의 점수보다 낮다면 패배한다" {
-        val dealer = Dealer()
-        dealer.init(initCard(spadeKing(), spadeFour())) // k(10) + 4 = 14
-        val player = player("test")
-        player.init(initCard(heartQueen(), heartTwo())) // q(10) + 2 = 12
-        player.match(dealer).matchResultType shouldBe MatchResultType.LOSE
-    }
-
-    "플레이어의 점수가 딜러의 점수와 같다면 무승부다" {
-        val dealer = Dealer()
-        dealer.init(initCard(spadeQueen(), spadeFour())) // q(10) + 4 = 14
-        val player = player("test")
-        player.init(initCard(heartQueen(), heartFour())) // q(10) + 4 = 14
-        player.match(dealer).matchResultType shouldBe MatchResultType.TIE
-    }
 })
